Create simplearraysum.py
Complete the simpleArraySum function in the editor below. It must return the sum of the array elements as an integer.

simpleArraySum has the following parameter(s):

ar: an array of integers

The first line contains an integer,n , denoting the size of the array.

The second line contains n space-separated integers representing the array's elements.

Print the sum of the array's elements as a single integer.